refactor: 💡 move model hook to session-recordings/index route
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-15130
