+++ ui/admin/app/routes/scopes/scope/session-recordings/index.js
@@ -4,5 +4,68 @@
  */
 
 import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
+import orderBy from 'lodash/orderBy';
 
-export default class ScopesScopeSessionRecordingsIndexRoute extends Route {}
+export default class ScopesScopeSessionRecordingsIndexRoute extends Route {
+  // =services
+  @service store;
+  @service router;
+  @service can;
+
+  // =attributes
+
+  queryParams = {
+    page: {
+      refreshModel: true,
+    },
+    pageSize: {
+      refreshModel: true,
+    },
+  };
+
+  /**
+   * Load all session recordings.
+   * @return {{sessionRecordings: Array, storageBuckets: Array}}
+   */
+  async model() {
+    let storageBuckets;
+
+    const scope = this.modelFor('scopes.scope');
+    const { id: scope_id } = scope;
+
+    if (
+      this.can.can('list scope', scope, {
+        collection: 'session-recordings',
+      })
+    ) {
+      const sessionRecordings = await this.store.query('session-recording', {
+        scope_id,
+        recursive: true,
+      });
+
+      // Sort sessions by created time descending (newest on top)
+      const sortedSessionRecordings = orderBy(
+        sessionRecordings,
+        'created_time',
+        'desc',
+      );
+
+      // Storage buckets could fail for a number of reasons, including that
+      // the user isn't authorized to access them.
+      try {
+        storageBuckets = await this.store.query('storage-bucket', {
+          scope_id,
+          recursive: true,
+        });
+      } catch (e) {
+        // no op
+      }
+
+      return {
+        sessionRecordings: sortedSessionRecordings,
+        storageBuckets,
+      };
+    }
+  }
+}
